+++ src/main/kotlin/dev/woytkowiak/ci/helper/ci/types/Ci3LibraryTypeProvider.kt
@@ -46,7 +46,9 @@ class Ci3LibraryTypeProvider : PhpTypeProvider4 {
         "xmlrpcs" to "CI_Xmlrpcs",
         "email" to "CI_Email",
         "typography" to "CI_Typography",
-        "uri" to "CI_URI"
+        "uri" to "CI_URI",
+        "security" to "CI_Security",
+        "form_validation" to "CI_Form_validation"
     )
 
     override fun getType(element: PsiElement): PhpType? {